+++ src/components/pages/apps/matcalc/apps/heatmap/heatmap-svg.tsx
@@ -225,16 +225,16 @@ function HeatMapSvgContent({ scale = 1 }: IProps) {
       ])
     )
 
-    const unadjustedInnerWidth = colLeaves.length * blockSize.w
+    const ungappedInnerWidth = colLeaves.length * blockSize.w
 
     const innerWidth =
-      unadjustedInnerWidth +
+      ungappedInnerWidth +
       displayOptions.gaps.cols.size * displayOptions.gaps.cols.indexes.length
 
-    const unadjustedInnerHeight = rowLeaves.length * blockSize.h
+    const ungappedInnerHeight = rowLeaves.length * blockSize.h
 
     const innerHeight =
-      unadjustedInnerHeight +
+      ungappedInnerHeight +
       displayOptions.gaps.rows.size * displayOptions.gaps.rows.indexes.length
 
     const width = innerWidth + margin.left + margin.right
@@ -315,7 +315,7 @@ function HeatMapSvgContent({ scale = 1 }: IProps) {
           displayOptions.tree.col.position === 'top' && (
             <ColTreeTopSvg
               tree={cf.colTree}
-              width={unadjustedInnerWidth}
+              width={ungappedInnerWidth}
               gaps={xgaps}
               height={displayOptions.tree.col.width}
               props={displayOptions}
@@ -366,7 +366,7 @@ function HeatMapSvgContent({ scale = 1 }: IProps) {
             <RowTreeSvg
               tree={cf.rowTree}
               gaps={ygaps}
-              width={unadjustedInnerHeight}
+              width={ungappedInnerHeight}
               height={displayOptions.tree.row.width}
               mode="left"
               props={displayOptions}
@@ -380,7 +380,7 @@ function HeatMapSvgContent({ scale = 1 }: IProps) {
             <RowTreeSvg
               tree={cf.rowTree}
               gaps={ygaps}
-              width={unadjustedInnerHeight}
+              width={ungappedInnerHeight}
               height={displayOptions.tree.row.width}
               mode="right"
               props={displayOptions}