+++ src/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Support for upper triangular display of heatmaps when square.
 - Bug fixes and updates.
 - Added back abilty for user to specify their own colors for GSEA gene hits.
 - Added simple text viewer app.